+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Respect DBusMenu `recursionDepth` when building menu layouts
 - Include `children-display` for submenu items in DBusMenu properties
 - Return all items for empty DBusMenu `GetGroupProperties` calls
+- Correct DBusMenu `AboutToShow` and `AboutToShowGroup` argument handling
 
 # 0.3.4 (2026-04-10)